mongodb/confluent-mdb-genai/semantic_filter/requirements.txt
requests
pymongo
boto3
langchain
langchain-community
wheel
kafka-python